Proper destruction of apic_t + runtime allocation example
The destruction of apic_t is now possible with the apic_t delete operator I added (also moved 'new' operator to apic_t) which will also free the class instance if it was runtime allocated. The unmapping of the mapped apic physical address was added to the xapic_t destructor.

The syntax for specifying compile time / runtime allocation has changed now, to just requiring APIC_RUNTIME_INSTANCE_ALLOCATION to be defined if runtime allocation is desired.

The runtime allocation example is a windows kernel mode driver which just defines APIC_RUNTIME_INSTANCE_ALLOCATION, and implements the routines allocate_memory and free_memory.