Fixed bug in apic_t::create_instance
Prevent numerical overflow - when apic_t::create_instance is called with static allocation enabled, it only allows one static instance, but the way we were checking this was bound to overflow after 256 attempts. By not incrementing it each time, but instead setting it to a constant value, we avoid this overflow.